tg_bot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pyTelegramBotAPI==4.19.1
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
telebot==0.0.5
urllib3==2.2.1
yarl==1.9.4
